Start refactoring modules
