+++ python/demo/stokes-taylor-hood/demo_stokes-taylor-hood.py
@@ -181,14 +181,14 @@ a_p = [[a[0][0], None],
 # Dirichlet boundary conditions are zeroed and a value of 1 is set on
 # the diagonal.
 
-A = dolfinx.fem.assemble_matrix_nest(a, bcs=bcs)
+A = dolfinx.fem.petsc.assemble_matrix_nest(a, bcs=bcs)
 A.assemble()
 
 # We create a nested matrix `P` to use as the preconditioner. The
 # top-left block of `P` is shared with the top-left block of `A`. The
 # bottom-right diagonal entry is assembled from the form `a_p11`:
 
-P11 = dolfinx.fem.assemble_matrix(a_p11, [])
+P11 = dolfinx.fem.petsc.assemble_matrix(a_p11, [])
 P = PETSc.Mat().createNest([[A.getNestSubMatrix(0, 0), None], [None, P11]])
 P.assemble()
 
@@ -196,10 +196,10 @@ P.assemble()
 # account for non-homogeneous Dirichlet boundary conditions:
 
 # +
-b = dolfinx.fem.assemble.assemble_vector_nest(L)
+b = dolfinx.fem.petsc.assemble_vector_nest(L)
 
 # Modify ('lift') the RHS for Dirichlet boundary conditions
-dolfinx.fem.assemble.apply_lifting_nest(b, a, bcs=bcs)
+dolfinx.fem.petsc.apply_lifting_nest(b, a, bcs=bcs)
 
 # Sum contributions from ghost entries on the owner
 for b_sub in b.getNestSubVecs():
@@ -207,7 +207,7 @@ for b_sub in b.getNestSubVecs():
 
 # Set Dirichlet boundary condition values in the RHS
 bcs0 = fem.bcs_by_block(extract_function_spaces(L), bcs)
-dolfinx.fem.assemble.set_bc_nest(b, bcs0)
+dolfinx.fem.petsc.set_bc_nest(b, bcs0)
 # -
 
 # Ths pressure field for this problem is determined only up to a
@@ -217,7 +217,7 @@ dolfinx.fem.assemble.set_bc_nest(b, bcs0)
 
 # +
 # Create nullspace vector
-null_vec = dolfinx.fem.create_vector_nest(L)
+null_vec = dolfinx.fem.petsc.create_vector_nest(L)
 
 # Set velocity part to zero and the pressure part to a non-zero constant
 null_vecs = null_vec.getNestSubVecs()
@@ -300,11 +300,11 @@ with XDMFFile(MPI.COMM_WORLD, "pressure.xdmf", "w") as pfile_xdmf:
 # matrices and iterative solvers.
 
 # +
-A = dolfinx.fem.assemble_matrix_block(a, bcs=bcs)
+A = dolfinx.fem.petsc.assemble_matrix_block(a, bcs=bcs)
 A.assemble()
-P = dolfinx.fem.assemble_matrix_block(a_p, bcs=bcs)
+P = dolfinx.fem.petsc.assemble_matrix_block(a_p, bcs=bcs)
 P.assemble()
-b = dolfinx.fem.assemble.assemble_vector_block(L, a, bcs=bcs)
+b = dolfinx.fem.petsc.assemble_vector_block(L, a, bcs=bcs)
 
 # Set near null space for pressure
 null_vec = A.createVecLeft()
@@ -454,15 +454,15 @@ L = form(inner(f, v) * dx)
 
 
 # Assemble LHS matrix and RHS vector
-A = dolfinx.fem.assemble_matrix(a, bcs=bcs)
+A = dolfinx.fem.petsc.assemble_matrix(a, bcs=bcs)
 A.assemble()
-b = dolfinx.fem.assemble.assemble_vector(L)
+b = dolfinx.fem.petsc.assemble_vector(L)
 
-dolfinx.fem.assemble.apply_lifting(b, [a], bcs=[bcs])
+dolfinx.fem.petsc.apply_lifting(b, [a], bcs=[bcs])
 b.ghostUpdate(addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE)
 
 # Set Dirichlet boundary condition values in the RHS
-dolfinx.fem.assemble.set_bc(b, bcs)
+dolfinx.fem.petsc.set_bc(b, bcs)
 
 # Create and configure solver
 ksp = PETSc.KSP().create(msh.comm)
