+++ python/dolfinx/fem/__init__.py
@@ -13,15 +13,17 @@ from dolfinx.cpp.fem import Form_float64 as _FormReal
 from dolfinx.cpp.fem import transpose_dofmap  # noqa
 from dolfinx.cpp.fem import IntegralType
 from dolfinx.cpp.fem import create_sparsity_pattern as _create_sparsity_pattern
-from dolfinx.fem.assemble import (apply_lifting, apply_lifting_nest,
-                                  assemble_matrix, assemble_matrix_block,
-                                  assemble_matrix_nest, assemble_scalar,
-                                  assemble_vector, assemble_vector_block,
-                                  assemble_vector_nest, create_matrix,
-                                  create_matrix_block, create_matrix_nest,
-                                  create_vector, create_vector_block,
-                                  create_vector_nest, pack_coefficients,
-                                  pack_constants, set_bc, set_bc_nest)
+# from dolfinx.fem.assemble import (apply_lifting, apply_lifting_nest,
+#                                   assemble_matrix, assemble_matrix_block,
+#                                   assemble_matrix_nest, assemble_scalar,
+#                                   assemble_vector, assemble_vector_block,
+#                                   assemble_vector_nest, create_matrix,
+#                                   create_matrix_block, create_matrix_nest,
+#                                   create_vector, create_vector_block,
+#                                   create_vector_nest, pack_coefficients,
+#                                   pack_constants, set_bc, set_bc_nest)
+from dolfinx.fem.assemble import (assemble_scalar, pack_coefficients,
+                                  pack_constants)
 from dolfinx.fem.bcs import (DirichletBCMetaClass, bcs_by_block, dirichletbc,
                              locate_dofs_geometrical, locate_dofs_topological)
 from dolfinx.fem.dofmap import DofMap
@@ -31,6 +33,7 @@ from dolfinx.fem.function import (Constant, Expression, Function,
                                   FunctionSpace, TensorFunctionSpace,
                                   VectorFunctionSpace)
 from dolfinx.fem.problem import LinearProblem, NonlinearProblem
+from dolfinx.fem import petsc
 
 
 def create_sparsity_pattern(a: typing.Union[_FormComplex, _FormReal]):
@@ -46,12 +49,10 @@ __all__ = [
     "Constant", "Expression", "Function",
     "FunctionSpace", "TensorFunctionSpace",
     "VectorFunctionSpace",
-    "create_vector", "create_vector_block", "create_vector_nest",
-    "create_matrix", "create_matrix_block", "create_matrix_nest",
-    "apply_lifting", "apply_lifting_nest", "assemble_scalar", "assemble_vector",
-    "assemble_vector_block", "assemble_vector_nest",
-    "assemble_matrix_block", "assemble_matrix_nest",
-    "assemble_matrix", "set_bc", "set_bc_nest",
+    # "create_vector", "create_vector_block", "create_vector_nest",
+    # "create_matrix", "create_matrix_block", "create_matrix_nest",
+    # "apply_lifting", "apply_lifting_nest", "assemble_scalar", "assemble_vector",
+    # "assemble_matrix", "set_bc", "set_bc_nest",
     "DirichletBCMetaClass", "dirichletbc", "bcs_by_block", "DofMap", "FormMetaClass", "form", "IntegralType",
     "adjoint", "LinearProblem", "locate_dofs_geometrical", "locate_dofs_topological",
-    "NonlinearProblem", "pack_coefficients", "pack_constants", "extract_function_spaces"]
+    "NonlinearProblem", "pack_coefficients", "pack_constants", "extract_function_spaces", "petsc"]
