+++ python/dolfinx/nls.py
@@ -29,9 +29,9 @@ class NewtonSolver(_cpp.nls.NewtonSolver):
 
         # Create matrix and vector to be used for assembly
         # of the non-linear problem
-        self._A = fem.create_matrix(problem.a)
+        self._A = fem.petsc.create_matrix(problem.a)
         self.setJ(problem.J, self._A)
-        self._b = fem.create_vector(problem.L)
+        self._b = fem.petsc.create_vector(problem.L)
         self.setF(problem.F, self._b)
         self.set_form(problem.form)
