+++ python/test/unit/fem/test_bcs.py
@@ -9,10 +9,10 @@ import pytest
 
 import ufl
 from dolfinx.fem import (Constant, Function, FunctionSpace,
-                         VectorFunctionSpace, apply_lifting, assemble_matrix,
-                         assemble_vector, create_matrix, create_vector,
-                         dirichletbc, form, locate_dofs_geometrical,
-                         locate_dofs_topological, set_bc)
+                         VectorFunctionSpace, dirichletbc, form,
+                         locate_dofs_geometrical, locate_dofs_topological)
+from dolfinx.fem.petsc import (apply_lifting, assemble_matrix, assemble_vector,
+                               create_matrix, create_vector, set_bc)
 from dolfinx.mesh import (CellType, create_unit_cube, create_unit_square,
                           locate_entities_boundary)
 from ufl import dx, inner
