+++ python/test/unit/fem/test_custom_assembler.py
@@ -22,8 +22,8 @@ import pytest
 import dolfinx
 import dolfinx.pkgconfig
 import ufl
-from dolfinx.fem import (Function, FunctionSpace, assemble_matrix, form,
-                         transpose_dofmap)
+from dolfinx.fem import Function, FunctionSpace, form, transpose_dofmap
+from dolfinx.fem.petsc import assemble_matrix
 from dolfinx.mesh import create_unit_square
 from ufl import dx, inner
