+++ python/test/unit/fem/test_fem_pipeline.py
@@ -10,9 +10,10 @@ import pytest
 
 import ufl
 from dolfinx.fem import (Function, FunctionSpace, VectorFunctionSpace,
-                         apply_lifting, assemble_matrix, assemble_scalar,
-                         assemble_vector, dirichletbc, form,
-                         locate_dofs_topological, set_bc)
+                         assemble_scalar, dirichletbc, form,
+                         locate_dofs_topological)
+from dolfinx.fem.petsc import (apply_lifting, assemble_matrix, assemble_vector,
+                               set_bc)
 from dolfinx.io import XDMFFile
 from dolfinx.mesh import (CellType, compute_boundary_facets, create_rectangle,
                           create_unit_cube, create_unit_square,
