+++ python/test/unit/fem/test_vector_assemble.py
@@ -7,7 +7,8 @@
 
 
 import ufl
-from dolfinx.fem import VectorFunctionSpace, assemble_matrix, form
+from dolfinx.fem import VectorFunctionSpace, form
+from dolfinx.fem.petsc import assemble_matrix
 from dolfinx.mesh import create_unit_square
 
 from mpi4py import MPI
